+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### Bugfix
 - ODPresentation : Exclude SVM files for reader - @Progi1984 GH-141
+- PowerPoint2007 Writer : Bugfix for opening PPTX on Mac - @thsteinmetz GH-89
 
 ### Changes
 - PhpOffice\PhpPowerpoint\Style\Alignment::setLevel can now be defined great than 8 - @Progi1984 GH-141